Enhance rendering and layering for native indicators
- Introduced a backdrop canvas for session highlights and gridlines, ensuring they render below all other layers.
- Updated the stacking logic to maintain the backdrop's position in the rendering order, allowing for proper visibility of indicators and gridlines.
- Removed redundant grid and highlight drawing logic from the Canvas2dBackend and WebGL2Backend, centralizing it in the new BackdropRenderer.
- Adjusted documentation to reflect changes in layer management and rendering behavior.
- Added tests to verify the integration of the backdrop canvas in the rendering pipeline.
